Add files via upload
It's a number guessing game developed in java. That is creating a one random number between 1 to 100 number. that is compare to user guessed number that is higher,less or not equal to and also equal to. and  it's give to hint to the user to guess higher or lower than that no.